musical-practice-companion/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
music21>=9.1.0
opencv-python>=4.8.0
pytesseract>=0.3.10
numpy>=1.24.0
pillow>=10.0.0
python-dotenv>=1.0.0
librosa>=0.10.0
scipy>=1.11.0
matplotlib>=3.8.0
PyMuPDF>=1.23.0
pdf2image>=1.16.0
